+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/MessageListState.kt
@@ -19,7 +19,6 @@ package io.getstream.chat.android.ui.utils.extensions
 import com.getstream.sdk.chat.adapter.MessageListItem
 import com.getstream.sdk.chat.view.messages.MessageListItemWrapper
 import io.getstream.chat.android.common.messagelist.MessageListState
-import io.getstream.chat.android.common.model.messsagelist.TypingItem
 
 /**
  * Converts the common [MessageListState] to ui-components [MessageListItemWrapper].
@@ -27,7 +26,7 @@ import io.getstream.chat.android.common.model.messsagelist.TypingItem
  * @return [MessageListItemWrapper] derived from [MessageListState].
  */
 public fun MessageListState.toMessageListItemWrapper(): MessageListItemWrapper {
-    var messagesList: List<MessageListItem> = messages.map { it.toUiMessageListItem() }
+    var messagesList: List<MessageListItem> = messageItems.map { it.toUiMessageListItem() }
 
     if (isLoadingOlderMessages) messagesList = messagesList + listOf(MessageListItem.LoadingMoreIndicatorItem)
     if (isLoadingNewerMessages) messagesList = listOf(MessageListItem.LoadingMoreIndicatorItem) + messagesList